Split tests for 8.4 and below
