+++ tests/PHPStan/Rules/Arrays/data/internal-classes-overload-offset-access-invalid-php84.php
@@ -1,9 +1,9 @@
-<?php
+<?php // lint >= 8.4
 /**
  * All of these offset accesses are invalid
  * ++ and -- are also disallowed in general are they operate "by ref"
  */
-namespace InternalClassesOverloadOffsetAccessInvalid;
+namespace InternalClassesOverloadOffsetAccessInvalid\Php84;
 
 function test1(\DOMNamedNodeMap $v): void
 {
@@ -15,7 +15,47 @@ function test1(\DOMNamedNodeMap $v): void
 	var_dump($r1);
 }
 
+function test2(\Dom\NamedNodeMap $v): void
+{
+	$v[] = 'append';
+	$v[0] = 'update';
+	$v[0] .= ' and again';
+	$r1 = &$v[0];
+	unset($v[0]);
+	var_dump($r1);
+}
+
 function test3(\DOMNodeList $v): void
+{
+	$v[] = 'append';
+	$v[0] = 'update';
+	$v[0] .= ' and again';
+	$r1 = &$v[0];
+	unset($v[0]);
+	var_dump($r1);
+}
+
+function test4(\Dom\NodeList $v): void
+{
+	$v[] = 'append';
+	$v[0] = 'update';
+	$v[0] .= ' and again';
+	$r1 = &$v[0];
+	unset($v[0]);
+	var_dump($r1);
+}
+
+function test5(\Dom\HTMLCollection $v): void
+{
+	$v[] = 'append';
+	$v[0] = 'update';
+	$v[0] .= ' and again';
+	$r1 = &$v[0];
+	unset($v[0]);
+	var_dump($r1);
+}
+
+function test6(\Dom\DtdNamedNodeMap $v): void
 {
 	$v[] = 'append';
 	$v[0] = 'update';
@@ -49,4 +89,3 @@ function test8(\ResourceBundle $v): void
 	var_dump($v['name']);
 	var_dump($v[0]);
 }
-
